Transaction ID moved from type to value
