+++ lib/domain/src/lib/transaction/type/index.ts
@@ -1,2 +1,1 @@
 export * from './transaction.js';
-export * from './transaction-id.js';
