--- lib/domain/src/lib/transaction/type/transaction-id.ts
@@ -1,17 +1,0 @@
-import { Schema } from "effect";
-import { randomUUID } from "node:crypto";
-
-export const TransactionId = Schema.UUID.pipe(
-    Schema.brand("TransactionId"),
-    Schema.annotations({
-        identifier: "TransactionId",
-    })
-)
-
-export type TransactionId = typeof TransactionId.Type;
-
-export type EncodedTransactionId = typeof TransactionId.Encoded;
-
-export const makeTransactionId = (id: string) => TransactionId.make(id);
-
-export const generateTransactionId = () => TransactionId.make(randomUUID());
