+++ lib/domain/src/lib/transaction/value/index.ts
@@ -1,0 +1,1 @@
+export * from './transaction-id.js';
