+++ lib/domain/src/lib/transaction/value/transaction-id.ts
@@ -1,41 +1,17 @@
-import { Schema } from 'effect';
+import { Schema } from "effect";
+import { randomUUID } from "node:crypto";
 
 export const TransactionId = Schema.UUID.pipe(
-  Schema.brand('TransactionId'),
-  Schema.annotations({
-    title: 'Transaction ID',
-    description: 'An unique transaction identifier in UUID format',
-  })
-);
+    Schema.brand("TransactionId"),
+    Schema.annotations({
+        identifier: "TransactionId",
+    })
+)
 
 export type TransactionId = typeof TransactionId.Type;
 
 export type EncodedTransactionId = typeof TransactionId.Encoded;
 
-/**
- *
- * Converts a raw value to a valid domain value
- *
- * @param encoded
- * @returns
- */
-export const fromTransactionId = (
-  encoded: EncodedTransactionId
-): TransactionId => Schema.decodeSync(TransactionId)(encoded);
+export const makeTransactionId = (id: string) => TransactionId.make(id);
 
-/**
- *
- * Converts a domain value to raw value
- *
- * @param decoded
- * @returns
- */
-export const toTransactionId = (decoded: TransactionId): EncodedTransactionId =>
-  Schema.encodeSync(TransactionId)(decoded);
-
-/**
- * Function to assert whether a value conforms to a given type
- */
-export const assertsTransactionId: Schema.Schema.ToAsserts<
-  typeof TransactionId
-> = Schema.asserts(TransactionId);
+export const generateTransactionId = () => TransactionId.make(randomUUID());
